fix: added loop logic to `time entries` in .pbit
This commit has now changed the Template file to include logic in the
Time Entries Query that will loop for as many pages of time entries are
available.

To achieve this I had to extract out the logic to get the time entries
into a separate function and call that within the `time entries` main
query.